Refactored out ID validating section
